+++ apps/ios-mockups/components/examples/texts.tsx
@@ -2,7 +2,7 @@ import { Iphone } from "@zuude-ui/ios-mockups";
 
 export default function Texts() {
   return (
-    <Iphone className="[--canvas-color:white]">
+    <Iphone className="[--screen-color:white]">
       <div className="flex flex-col gap-4 text-black pt-[var(--top-safe-area)] px-4">
         <h1 className="text-2xl font-bold">Hello World</h1>
         <p className="text-sm">