+++ packages/ui/src/provider/app-wrapper.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Navbar } from "../components/navbar";
 import { Providers } from "./next-theme";
 import { Toaster } from "../components/sonner";
+import { ReactScan } from "../components/react-scan";
 
 interface Props {
   children: React.ReactNode;
@@ -13,6 +14,7 @@ export function AppWrapper({ children }: Props) {
       <Navbar />
       {children}
       <Toaster />
+      <ReactScan />
     </Providers>
   );
 }