+++ apps/ios-mockups/CHANGELOG.md
@@ -1,5 +1,19 @@
 # ios-mockups
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/ios-mockups@0.0.8
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/ios-mockups@0.0.7
+
 ## 0.0.5
 
 ### Patch Changes